api/requirements.txt
flask==2.3.3
flask-cors==4.0.0
pymysql==1.1.0
python-dotenv==1.0.0




